[DEV-3583] changes after review
