+++ src/shared/services/dfx-cron.service.ts
@@ -2,11 +2,11 @@ import { Injectable, OnModuleInit } from '@nestjs/common';
 import { DiscoveryService, MetadataScanner } from '@nestjs/core';
 import { CronExpression, SchedulerRegistry } from '@nestjs/schedule';
 import { CronJob } from 'cron';
-import { DfxLogger } from 'src/shared/services/dfx-logger';
 import { DisabledProcess } from 'src/shared/services/process.service';
-import { DFX_CRONJOB_PARAMS, DfxCronParams } from 'src/shared/utils/cron';
+import { DFX_CRONJOB_PARAMS, DfxCronExpression, DfxCronParams } from 'src/shared/utils/cron';
 import { LockClass } from 'src/shared/utils/lock';
 import { Util } from 'src/shared/utils/util';
+import { CustomCronExpression } from '../utils/custom-cron-expression';
 
 interface CronJobData {
   instance: object;
@@ -17,8 +17,6 @@ interface CronJobData {
 
 @Injectable()
 export class DfxCronService implements OnModuleInit {
-  private readonly logger: DfxLogger = new DfxLogger(DfxCronService);
-
   constructor(
     private readonly discovery: DiscoveryService,
     private readonly metadataScanner: MetadataScanner,
@@ -51,22 +49,50 @@ export class DfxCronService implements OnModuleInit {
   private addCronJob(data: CronJobData) {
     const lock = LockClass.create(data.params.timeout ?? Infinity);
 
-    const cronJob = new CronJob(data.params.expression, () => lock(this.wrapFunctionInTryCatchBlocks(data)));
+    const cronJob = new CronJob(data.params.expression, () => lock(this.wrapFunction(data)));
     const cronJobName = `${data.instance.constructor.name}::${data.methodName}`;
 
     this.schedulerRegisty.addCronJob(cronJobName, cronJob);
     cronJob.start();
-
-    this.logger.info(`CronJob ${cronJobName} started`);
   }
 
-  private wrapFunctionInTryCatchBlocks(data: CronJobData) {
+  private wrapFunction(data: CronJobData) {
     return async (...args: any) => {
       if (data.params.process && DisabledProcess(data.params.process)) return;
 
-      if (data.params.useDelay ?? true) await Util.cronJobDelay(data.params.expression as CronExpression);
+      if (data.params.useDelay ?? true) await this.cronJobDelay(data.params.expression as CronExpression);
 
       await data.methodRef.apply(data.instance, args);
     };
+  }
+
+  private async cronJobDelay(expression: DfxCronExpression): Promise<void> {
+    const random = Math.random() * 1000;
+
+    switch (expression) {
+      case CronExpression.EVERY_10_SECONDS:
+        return Util.delay(random * 5); // 0 .. 5 sec
+
+      case CustomCronExpression.EVERY_15_SECONDS:
+        return Util.delay(random * 5); // 0 .. 5 sec
+
+      case CronExpression.EVERY_30_SECONDS:
+        return Util.delay(random * 10); // 0 .. 10 sec
+
+      case CronExpression.EVERY_MINUTE:
+        return Util.delay(random * 10); // 0 .. 10 sec
+
+      case CronExpression.EVERY_5_MINUTES:
+        return Util.delay(random * 30); // 0 .. 30 sec
+
+      case CronExpression.EVERY_10_MINUTES:
+        return Util.delay(random * 30); // 0 .. 30 sec
+
+      case CustomCronExpression.EVERY_15_MINUTES:
+        return Util.delay(random * 30); // 0 .. 30 sec
+
+      case CronExpression.EVERY_HOUR:
+        return Util.delay(random * 60); // 0 .. 60 sec
+    }
   }
 }
